move test logic into classes
