+++ tests/helper/exception.py
@@ -1,0 +1,5 @@
+class NotPartOfFeatureError(Exception):
+    pass
+
+class TestFailed(Exception):
+    pass
